Interactive and Analysis Script Utility Functions.

This package consists of aliases, functions, and graphical themes to improve your efficiency and workflow.

Installation

Install this package:

install.packages("devtools")
devtools::install_github("dgabbe/dgutils")

Aliases

Create some bash equivalents in aliases.R

Functions

Utility/convenience functions are defined in their own R file named for the function.

Graphical Themes

This section is graphics programming constructs. Since R has a number of different packages, there's a separate file for each one. Only ggplot is populated.

<em>graphics-ggplot.R</em>

The ggplot2 package is somewhat unique from other packages with heavy use of the "+" continuation operator and the huge number of objects and functions needed to generate a graph.